functions/import_loto_results_to_bq/requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.27.1
functions-framework==3.8.2
